requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
tensorflow-cpu==2.18.0
torch==2.2.1
torchvision==0.17.1
pillow==10.2.0
python-multipart==0.0.9
pydantic==2.6.4
numpy==1.26.4
psutil==5.9.8